Fix: Linter errors (#6684)
* Clean up utilities/images.ts

* More random linter fixes

* More random linter fixes 2
