+++ packages/manager/src/components/PasswordInput/HideShowText.tsx
@@ -5,7 +5,7 @@ import { TextFieldProps } from 'src/components/core/TextField';
 import TextField from '../TextField';
 
 interface State {
-  hidden: Boolean;
+  hidden: boolean;
 }
 
 type Props = TextFieldProps & {
