+++ packages/manager/src/features/linodes/CloneLanding/Details.tsx
@@ -199,7 +199,6 @@ export const Configs: React.FC<Props> = props => {
                   onClick={() => handleToggleConfig(eachConfig.id)}
                   className={classes.closeIcon}
                   data-qa-inline-delete
-                  role="button"
                 >
                   <Close />
                 </button>
@@ -231,7 +230,6 @@ export const Configs: React.FC<Props> = props => {
                 onClick={() => handleToggleDisk(eachDisk.id)}
                 className={classes.closeIcon}
                 data-qa-inline-delete
-                role="button"
               >
                 <Close />
               </button>
