+++ packages/manager/src/features/linodes/LinodesLanding/IPAddress.tsx
@@ -144,7 +144,7 @@ export class IPAddress extends React.Component<Props & WithStyles<CSSClasses>> {
     );
   };
 
-  renderIP = (ip: string, copyRight?: Boolean, key?: number) => {
+  renderIP = (ip: string, copyRight?: boolean, key?: number) => {
     const { classes, showAll } = this.props;
     return (
       <div
