+++ packages/manager/src/utilities/haveAnyBeenModified.ts
@@ -1,7 +1,7 @@
 import { equals } from 'ramda';
 
 let haveAnyBeenModified: <T>(objA: T, objB: T, keys: (keyof T)[]) => boolean;
-
+// eslint-disable-next-line
 haveAnyBeenModified = (state, nextState, keys) => {
   let idx = 0;
   const len = keys.length;
